dodano pobieranie informacji z zlecenia za pomocą openai

+++ PotoDocs.API/PotoDocs.API/Models/InvoiceDto.cs
@@ -12,7 +12,7 @@ public class InvoiceDto
     public int InvoiceNumber { get; set; } // NUMER_FAKTURY
     public string CompanyName { get; set; } // NAZWA_FIRMY
     public string CompanyAddress { get; set; } // ADRES_FIRMY
-    public int CompanyNIP { get; set; } // NIP_FIRMY
+    public long CompanyNIP { get; set; } // NIP_FIRMY
     public DateTime SaleDate { get; set; } // DATA_SPRZEDAZY
     public DateTime IssueDate { get; set; } // DATA_WYSTAWIENIA
     public int PaymentDueDate { get; set; } // TERMIN_ZAPLATY
